FindSpam Refactor
Split out the rules, helpers and blacklists.
Start a folder structure.
