+++ Helpers/__init__.py
@@ -1,0 +1,3 @@
+"""
+The FindSpam rules and checks package
+"""
